refactor(tui): extract monolithic tui.go into focused modules
Split the large tui.go (~2800 lines) into smaller, single-responsibility files:
- content_render.go / content_line.go: message content rendering
- view_render.go: view layout and rendering
- input_layout.go / input_views.go: input component layout
- update.go: BubbleTea Update message handling
- approval.go: approval UI component
- history.go: session history navigation
- list_items.go: list item delegates
- pickers.go / perf.go: picker components and perf utilities
- app.go (channel_panel, manage_models, ssh_handlers): sub-components

This improves maintainability without changing any behavior.
